Move transmutation recipe registry to init package
There is no reason why it should be in the API.
