+++ src/main/java/jard/alchym/api/ingredient/IngredientGroup.java
@@ -3,10 +3,8 @@ package jard.alchym.api.ingredient;
 import com.google.common.collect.Lists;
 import jard.alchym.api.ingredient.impl.FluidVolumeIngredient;
 import jard.alchym.api.ingredient.impl.ItemStackIngredient;
-import jard.alchym.api.recipe.RecipeGroupAccessor;
-import jard.alchym.api.transmutation.TransmutationAction;
+import jard.alchym.init.RecipeGroupAccessor;
 import jard.alchym.api.transmutation.TransmutationInterface;
-import jard.alchym.api.transmutation.impl.DryTransmutationInterface;
 import alexiil.mc.lib.attributes.fluid.volume.FluidVolume;
 import net.minecraft.entity.ItemEntity;
 import net.minecraft.item.ItemStack;
